Generacion de boleta interna

+++ src/main/java/control_calificaciones/models/CalificacionH.java
@@ -18,7 +18,7 @@ public class CalificacionH {
     @EqualsAndHashCode.Include
     private Integer idCalificacion;
 
-    private Integer resultado;    
+    private Double resultado;    
 
     @ManyToOne(cascade =  CascadeType.PERSIST)
     @JoinColumn(name = "curp", referencedColumnName = "curp")
